+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - feature: install postman
 - feature: install qwen code
 - feature: install sourcegit
+- feature: install vs code insiders
 
 ## [1.4.0] - 2026-07-24